backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pydantic==2.7.0
pydantic-settings==2.2.1
python-multipart==0.0.9
jira==3.8.0
py-trello==0.19.0
azure-devops==7.1.0b4
msrest==0.7.1
python-gitlab==4.4.0
pandas==2.2.2
numpy==1.26.4
python-dateutil==2.9.0
statsmodels==0.14.2
apscheduler==3.10.4
aiofiles==23.2.1
httpx==0.27.0
cryptography==42.0.8
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-cov==5.0.0
